Cambios en base de datos

+++ ScripBaseDeDatos/Backend/PuraVidaStoreBK/PuraVidaStoreBK/Models/DTOS/ImpuestosDTO.cs
@@ -6,7 +6,7 @@
 
         public string? ImpDescripcion { get; set; }
 
-        public double? ImpPorcentaje { get; set; }
+        public decimal? ImpPorcentaje { get; set; }
 
         public bool? ImpActivo { get; set; }
     }
